Consolidate desktop UI tray coordination
